services/core/requirements/base.txt
alembic==1.16.4
asyncpg==0.30.0
dotenv==0.9.9
fastapi==0.116.1
httpx==0.28.1
pydantic-settings==2.10.1
SQLAlchemy==2.0.42
uvicorn[standard]==0.35.0
